Ensure callback is run with correct progressRunner
If a sdk load is already in progress and another request comes in,
the callbacks from the later request are added to the existing task.
Previously they would get run with the progressrunner of the original
task, which could (e.g. if the progressrunner is going to invokeLater)
behave differently than the callback expects.
In this change we keep the runner around along with the callback.

Change-Id: I6219db0818215f5b6a5ee02d5522b114b8495104