do not produce main split when warmswap is possible.
when in 23, and resources have changed but all the code changes
(if any) are compatible, do not produce the main split APK, the
resources.ap_ is enough.

remember that the main APK should be built when the next
incompatible change comes around.

finally purge resources entries once an incompatible change has
been detected.

Change-Id: If629dc5591f64e3a6fb1eb54f49e4416b7264221