Fixed realignment of streamed zips.
Streamed entries in zips usually have the CRC *after* the entry data.
When aligning the zip, we rewrite the entry and this was failing as
we know the CRC upfront so there is no need to put if after the entry
data (and have one CRC as zero before and the real one after).

This fix also has the advantage that, by removing the CRC section
after the entry, we may not need to touch any other entries in the
zip while realigning because the saved space from the end of the
entry can be used for aligning.

Change-Id: Ia897115f6c6afd72831de537ea31081624a11591