Updating RepositoryPackages in-place
Previously, when a reload was run, a new RepositoryPackages
was created. This meant that after the local load completed but before
the remote was started, the new local packages were only available in
the RepositoryPackage instance passed in to the repo update callback,
and not via any apis on RepoManager itself. Specifically, this meant
that after installing a new package, when SystemImageList tried to get
the new list of system images, the AndroidTargetManager and
SystemImageManager would pick up the previous local packages (since
they were created as part of the local complete callback), and the newly
installed package wouldn't be seen.

Modified such that the RepositoryPackages instance is updated in place.
Also changed RepositoryPackages so the computed new and updated package
lists as well as the UpdatablePackages are created lazily (and nulled
out when the local or remote packages change), since now the "normal"
complete load will update the local packages and then soon after the
remote packages--we don't want to recalculate these lists before the
second update happens.

Change-Id: I6eac484e0eb0126351a415a0b58f036123febfe6