196494: Fix quote stripping in StringFormatDetector
Currently we would strip quotes if the are at the beggining and the end
of a text block. In some cases, the quote at the end could be escaped
leaving the escape character and removing the quote incorrectly. Because
we leave the escape character, when concatenated with the next text
node, the first character of the next node will be escaped and therefore
the result will be wrong.
This CL changes the quote stripping to avoid those cases.

Change-Id: Ie7e41e674459db5c224740b6779d70072a985559