do not keep previous change records for deleted classes.
when changes are compatible, we keep a list of changed files to
process when an incompatible change comes around. However, there
is no need to keep the changed records for classes that have been
deleted in the current iteration.

Change-Id: I4e2b4bd35f6b451e0ec2df3ee3cde754fade738f
(cherry picked from commit 65a167af797994b1cebde0bb79100555bae907cf)