Added best/default deflate ratio choice.
If the new BestAndDefaultDeflateExecutorCompressor is used
as a compressor, both the best and deflate algorithms are
attempted to compress the file and the default is chosen
if above a certain ratio.

This allows choosing default compression when the
compression savings are not big but speeding up greatly
Play Store handling of APKs.

Change-Id: Ib6d5791a7c64b821101a0b55a1b98abe864bf970