Consider drawable, transition, xml to be ID generating.
Previously, a lot of code assumed that only menu and layout
were ID generating resource folder types.
In the case of drawable, only consider files that are XML.

Drawable XML may contain IDs for the layers, and the IDs
can then be used in Java by "findDrawableByLayerId" and by
"setDrawableByLayerId", etc.

Found this while comparing resource repos with resource
managers (managers used the ID_PROVIDING_RESOURCE_TYPES
list in AndroidCommonUtils to determine when to index IDs).
AndroidDrawableResourcesDomTest would fail if we converted
to resource repos without this.

Bug: b/25078759
Change-Id: I6dc3529bbb9914b110f1dca1193e9858d843abf4