Get rid of sdk common dependency in the api generator
When the api generator is used in platform builds, it depends on
the prebuilts/devtools/tools/lib/common.jar prebuilt -- which isn't
versioned. We could replace it with a reference to the offline
m2 repository, but the purpose of that one is for Studio's offline
repo which may change independently of the platform builds.

Turns out we only need two trivial things from common (the Pair
class and a trivial constant), so just inline those two things
instead to get rid of the whole dependency.

Change-Id: If711817546da8b0be9fa185a24af3682782991f5