Fix handling of private methods in find-super and overrides
The code to resolve super methods, and to handle overriding
methods (for annotations inheritance for example) was incorrectly
handling private methods. If a class has a private method, and
a subclass has a public method of the same signature, that
subclass method isn't "overriding" the super class method. However,
it *can* call super if they're in the same compilation unit.

Change-Id: Iebe40be9932fdc55a6abd73fc517ddf2eb3fbacf