wait until dexing is finished to add to context.
up to now, we assume the classes.dex is the only file resulting
from dexing a dependency but if the dependencies are really large,
there can be more than one with multidexing on. visit the directory
after the process is finished and all results.

Change-Id: I1f168e33485b0ae1201c156b525363db409520d0
(cherry picked from commit c6ac0e556a323b16624982a88b8291647304ccc8)