do not package classes in the APK as dex files.
The application classes used to be packaged as
dex files in the resulting APK. Now all application's
code will be packaged inside a zip file as a java
resource (root of the APK). The instant run runtime
will continue to be shipped as classes.dex in the main APK.
Upon initialization, the instant run runtime will recover
that zip file and extract all dex files to a location
used by the classloaders.

Change-Id: I7b0987768178abbb7140ef1f3a628bcfbaca0b02