Adding ability to fall back to previous schema versions
If e.g. studio is updated and gradle is not, studio might write out a
local package in a newer version of the schema than what gradle knows
about. In this case we want gradle to be able to do a best-effort parse
of the xml using its latest known schema versions.
This CL adds that functionality.

Change-Id: Id6f39034a1a15f3a64a571286d76525ac990bd94