do not publish instant run enabled manifest.xml
some plugins rely on the merged manifest to customize generated
code and up to now, they were pointed to the InstantRun
decorated version. With this change, the gradle-api pointer to the
merged manifest will continue pointing to the merged manifest
while we will use a different location for the instant run enabled
manifest (which we'll copy in the APK).

BUG=195033

Change-Id: I2e8760392030e42917437e699a424d74505f10c9