Suppress annotation tests in ExtractTest
The ExtractTest tests annotation extraction and includes a lot
of support annotation test cases which checks that the annotations
are correctly extracted. Some of these are not valid for the element
they are specified on; lint now checks for this, but that's not
important for the extraction test so just suppress these lint usage
warnings.

Change-Id: I660d05dc3db7eb6b70d34fa1cf33e94e2dd4dd55