Refactoring installers
Previously, there was no obvious way to extend the list of listeners
added to a package installer. In this change the concept of a listener
factory is added, which provides a list of listeners and can be
subclassed, where the subclasses can extend that list.

Previously, install and uninstall functionality were provided by the
same class (PackageInstaller). This was somewhat awkward, as they
operate on RemotePackages and LocalPackages respectively, and casting
was necessary in every case. Now Installers and Uninstallers are split
out into separate classes. InstallerFactories are also introduced, where
a single InstallerFactory can create either the Installer or
corresponding Uninstaller instance.

Change-Id: Ic629d20de05359ec36469cbfea2b72060765ac02