First version of native server running on device for Android Studio Profilers.
It is a skeleton for future development, validating the dev work flow and tool chain.

This particular change includes the functionality to read and parse CPU utilization info from /proc/stat. It includes a very naive server and a very naive sample client for illustration purpose.

Both the server.cpp and client.cpp will be very short-lived. The server (server.cpp) will likely be completely rewritten, to implement Android Studio Profiling Protocol such as the handshake, message types. When the server is rewritten, it is assured that the sample client needs to be rewritten, too.

The code to read /proc/stat may be more mature and is expected to stay.

Change-Id: I8cd301941425c6b6146f81cefaa1af114e5eb4cc