Make new lint PSI backend handle unresolved symbols
I ran lint on a large codebase of sample projects, many
older one with incomplete classpaths. This revealed
a number of places where ECJ leaves null or incomplete
fields that it should defensively check for. This won't
happen when lint is running within Gradle, since the
build would have failed before running lint if there are
incomplete paths, but this can happen when the lint
command line script is invoked directly on a tree.

Also found and fixed a few other bugs:
- a false positive in the commit-fragments detector
- a few missing language constructs converting from ECJ to PSI
  (assignments and method calls in the first for-loop clause, as well
  as nested annotations with array initializer syntax.
- the "findContinuedStatement" for continue statements (and
  findExitedStatement for break statements) had not been implemented;
  now they are.

Change-Id: Icf47927b5059117da1e51a47d1c4eaa317c6723f