NodeUtils: handle non-root xmlns:prefix=uri declarations
doc.createAttributeNs seems to expect a qualifiedName, so
pass in attr.getNodeName instead of attr.getLocalName.
Otherwise it will throw a NAMESPACE_ERR exception.

I think from the check:

(namespaceURI!=null &&
  namespaceURI.equals(NamespaceContext.XMLNS_URI) &&
  !qname.equals("xmlns"))

when there is no prefix in the qname, but the
attribute's namespace is the XMLNS_URI.

In the duplicated document, the new NS prefixes get
declared at the top level, so erase the local ones
in the duplicated document.

Bug: b/27827198
Change-Id: Ie184e4301e62d93993bdb5317f9dd10552f84249