remove unwanted trace and bumped up default dex processes.
we used to have 3 dex processes including the one for dependencies.jar,
but since we don't create the dependencies.jar any longer, the number
of processes went down to 2. Re-enabling 3 processes.

changed a warning when the dex processes settings
cannot be applied and give a hint on  how to solve the issue.

removed unwanted trace.

Change-Id: Ida5587a556c3501db36d65511b539238caaf7b59