Add a pre-compilation task to wipe out folders.
This pre-compilation task will wipe out the javac compilation output folder
when it is running. Its input are the generated source folders (where R.java
lives) and will be triggered anytime one of these generated source files has
changed.

This is a workaround for the incremental java compilation support not working
in Gradle when having more than one input folder.

Change-Id: Iecdba4fddd4c458c7ea395e2d5a36738273107e7