Add non continuous (stepped) line charts
In a continuous LineChart, a straight line is drawn between points.
For example, (x0, y0) and (x1, y1).

However, there are some scenarios in which a continuous LineChart doesn't
make sense. For instance, when we want to check how many CPUs were not
idle in a determined point in time. Non continuous line charts cover
these types of scenarios.

In a stepped LineChart, a line is drawn from (x0, y0) to (x0, y1)
and another one is drawn from (x0, y1) to (x1, y1).

Change-Id: I7d91f45b0a13353109ed79265108cc3fb962c62c