Handle exception thrown if SelectionComponent is not showing
If a component is not showing, calling getLocationOnScreen() will cause
an IllegalComponentStateException to be thrown. We should handle that in
getMouseLocation() method of SelectionComponent.

This CL also fixes the Cursor that shows up when selecting the
max (right) handle. It should also point to the right, not to the left.

Change-Id: I9144657f5f907cbe6fc55a2932f8dd22dda00223