Merge resources uses the new Aapt interface.
Updated MergeResources to use the new Aapt interface
instead of the PnhCruncher interfaces. AaptV1 uses those
interfaces when compiling resources, but its usage is
now hidden behind the generic Aapt interface. This
allows support for compiling resources with aapt2 as
soon as there is an implementation of the Aapt
interface that uses the aapt2 interface.

Change-Id: I737ee33c70f627bf181a8a2bb456b62739ccd1ca