Remove misleading annotations from BuildInfoLoaderTask.
The annotations don't work on fields, so are not performing the expected
validation.

For example, @InputFile @Optional is actually @Nullable at the point of task
execution, with the file validated to exist if the field is not null.  This
task incorrectly assumes it means that the file must be specified but is
allowed not to exist.

Change-Id: I292dfabea6fc315447814764c0259c06f86b2812