Consolidating Installer and Uninstaller interface
The patcher-based uninstaller will need to do some prep work which it makes sense
to have in a prepare phase, similar to installers. Additionally, it's weird that
the UI for installing and uninstalling is currently different (that is,
there's basically no UI for uninstalling).
This change makes both of those upcoming changes easier.

As a side effect, MavenInstallListner was rewritten to regenerate the
whole metadata file each time, rather than trying to update the existing
file based on the installed/uninstalled packages.

Change-Id: I1e722eaca93373a4cd11baada8798ec88a03c5ea