Changes ProfilerService initialization
This CL refactors the way the ProfilerService is initialized
instead of assuming there is an Application class, this plugin
looks for all known entry points (for now only Application and
Activity) and initializes our Service there.

This change also separates the transform specific behaviour
and creates a class to allow simple class transformations.

Network and Initialization are now separate MethodVisitors.

Deleted ProfilerServer and moved the ProfilerComponent hooks
to the Service.

Change-Id: I15c14d03e89c840e88cc80435d8b394657a3034b