Initial import of arsc blamer sources
This CL imports the source code from
https://github.com/google/android-arscblamer @ 2a6dab14 with one minor
change: the test apk file has been replaced (from Topeka.apk to
a simpler base.apk simply to reduce the size).

The current import places the source code into a separate folder
that is not marked as in source or resource root. Subsequent CLs will
clean this up and import just the necessary files in a form that is
compatible with the tools build.

Change-Id: I3d3413c2e1258ec98839a855a3c049c998904eda