Extract PackagingScope
Since packaging uses a combination of global, variant and variant output
scopes, I created a new interface and an implementation that delegates
to the appropriate level, instead of three interfaces. A different
implementation will be used in the context of an external build system
build.

I also moved ValidateSigningTask creation to TaskManager.

Change-Id: I7354f966529b32b012ff7a86bb8002c99e34687e