Change protobuf to not find zlib, but to use the compiled one.
protobuf_WITH_ZLIB=0 tells protobuf to not try to find zlib.
There is a way of changing it to point to where zlib is installed
but in our case we need to compile it from source.

Adding HAVE_ZLIB=1, makes protobuf compiling assuming zlib is
provided which enables compressed streams.

Change-Id: I34c19e38927713263a98d3a9a61ebd301dfdcd4a