Hooks up perfa with supportlib.
This CL adds perfa as a jni library of supportlib. This integration is
disabled by default as it needs gRPC to compile perfa.

Long term perfa will only be added as an agent, but before there is
support for that supportlib will initialize it instead. The new
directory "support" in perfa is the layer to interact with supportlib.

Change-Id: I81860e4744f29c9c9b8333ad71662d2cfa7b0881