Make sure ndk-build project that uses Google Test works appropriately
Google Test is part of the NDK. This doesn't add Google Test support to
external native build. It just makes sure that a project that  lready has 
Google Test in the Android.mk works correctly. 

This CL also causes .a targets to be emitted for ndk-build. This already
happens for CMake.

b.android.com/211504: ndk-build command line analyzer need to emit targets for -shared objects

Change-Id: I04b3f750df68f307c0caaa9dc2e2a3f3a701aed4