Rename Sampler <-> Collector
The current codebase has classes called "*Collector" and "*Sampler",
with samplers delegating to a list of collectors. Now, we flip this,
as Samplers should be responsible for pulling out small data samples,
and collectors should loop through a list of Samplers.

Also, simplified names by removing the redundant 'data' from them,
since it's implied that a sampler would be sampling data.

Change-Id: I33848cea85eb7c6c460d73a70c8ee8669a1e9ef0