Adds perfa to studio builds as a JNI component.
Embeds libperfa.so in supportlib.

When building for windows CMake is skipped. This CL also
changes gradle bundling tasks to include only what's available so
on windows builds there won't be any .so available.

This change also fixes gradle <-> cmake task dependencies

CMake is only run if either the toolchain files have changed or if
the cache is not present.

All ninja tasks are run unconditionally as ninja takes care of
dependencies itself and it's super fast.

Change-Id: I7d29977a1d0585cf2d3013db54a581a2aa75c4f2