Build system performance logging for the masses.
This builds on the RECORD_SPANS infrastructure to write an
AndroidStudioEvent proto for each build invocation containing
detailed performance information.

These are written to a file which studio will then pick up and upload.

This is only enabled if the user has opted in to sending usage
statistics in studio.

Change-Id: I95fa61094a852b93451f547b25aebb8758bc4f7d