New ByteBatcher utility class
We'll be sending data from Supportlib to perfa to perfd, which
crosses a Java -> C++ boundary and then jumps over a GRPC
boundary.

Doing this for each byte received would be overkill, so I introduce
a utility class which batches bytes and only triggers logic after a
threshold (currently defaulting to 1K) is passed.

Change-Id: If09cdf0341fd3ad8acae4e38b6393ed4dc691471