Add support for path operations
With tests!

Path operations are add, subtract, intersect and xor. They are fairly
common in complex shapes. Supporting them means that a decoded Path2D
only every contains a single sub-path (1 moveTo command). Paths could
in theory be merged but it requires complicated heuristics (2 paths
with an add op can only be merged if they are both using a non-zero
winding rule or if they don't overlap and they are even-odd).

Change-Id: I75ddb2f3a8f748d6e55ee419f0e63cb887595a9a